# src/spec/swagger.json
{"openapi":"3.0.0","paths":{"/health-check":{"get":{"operationId":"AppHealthCheck","parameters":[],"responses":{"200":{"description":""}}}},"/identity/config":{"get":{"operationId":"IdentityGetConfig","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigModel"}}}}},"tags":["identity"]}},"/identity/type":{"get":{"operationId":"IdentityGetType","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["identity"]}},"/identity/identifier":{"get":{"operationId":"IdentityGetIdentifier","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["identity"]}},"/identity/tokens":{"get":{"operationId":"IdentityGetTokens","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["identity"]}},"/signal/latest/{tokenPair}":{"get":{"operationId":"SignalCoreGetLatest","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignalModel"}}}}},"tags":["signal"]}},"/signal/latestPrice/{tokenPair}":{"get":{"operationId":"SignalCoreGetLatestAction","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["signal"]}},"/signal/latestTime/{tokenPair}":{"get":{"operationId":"SignalCoreGetLatestTime","parameters":[{"name":"tokenPair","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["signal"]}}},"info":{"title":"PriceEmitter","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"schemas":{"StringObject":{"type":"object","properties":{"value":{"type":"string"}},"required":["value"]},"ConfigModel":{"type":"object","properties":{"type":{"type":"string","default":"price"},"identifier":{"type":"string","default":"test"},"tokens":{"default":[],"type":"array","items":{"type":"string"}}},"required":["type","identifier","tokens"]},"SignalModel":{"type":"object","properties":{"tokenPair":{"type":"string"},"action":{"type":"number"},"timestamp":{"type":"integer","format":"int64"}},"required":["tokenPair","action","timestamp"]}}}}